project-oracle/project/requirements.txt
Django
Fabric
django-webpack-loader
djangorestframework
markdown
django-filter
pandas
numpy
plotly
pymining
nltk
scipy
jellyfish
